Import 2.1.13

+++ include/linux/kdev_t.h
@@ -62,13 +62,13 @@ aeb - 950811
    However, everything runs fine with an arbitrary kdev_t type. */
 
 #define MINORBITS	8
-#define MINORMASK	((1<<MINORBITS) - 1)
+#define MINORMASK	((1U << MINORBITS) - 1)
 
 typedef unsigned short kdev_t;
 
-#define MAJOR(dev)	((dev) >> MINORBITS)
-#define MINOR(dev)	((dev) & MINORMASK)
-#define HASHDEV(dev)	(dev)
+#define MAJOR(dev)	((unsigned int) ((dev) >> MINORBITS))
+#define MINOR(dev)	((unsigned int) ((dev) & MINORMASK))
+#define HASHDEV(dev)	((unsigned int) (dev))
 #define NODEV		0
 #define MKDEV(ma,mi)	(((ma) << MINORBITS) | (mi))
 #define B_FREE		0xffff		/* yuk */
